added migrations

+++ database/migrations/2022_07_11_121716_create_people_table.php
@@ -19,8 +19,7 @@ class CreatePeopleTable extends Migration
             $table->string("lastName");
             $table->date("birthDate");
             $table->string("department");
-            $table->float("maxAmountOfHolidays");
-
+            $table->integer("maxAmountOfHolidays");
             $table->timestamps();
         });
     }
